Start splitting game state out
This will be something settable in some kind of
UI, and then passed to the effect eval functions.
Helps keep things split up nicely, so that the
UI can just set data on the context object
and the eval functions can work on that instead
of trying to put data directly into something
that is holding parsed data from the JSON or
adding a bunch of parameters or other similarly
messy things.