Fix ahanbi ability 2; Start unit tests
Also start implementing a utility on the dmg
formula class to output everything that has
been changed from its base state. This will
allow for easier debugging/unit testing to
be able to easily see everything that has been
set on the formula. As well as being able to
verify that in code, such as making sure that
ONLY an attack modifier has been set on it and
nothing else.