Fix allowing partial condition triggers
This was revealed when testing Vagner's
ability 2, which increments the attack
modifier after every 5 powerflips. When setting
the state to have 11 powerflips instead of an
even number like 5 or 10, it was calculating
a 0.33 increase to attack instead of 0.3
due to not reaching the next activation point
of 0.5. This bug has been fixed.