Start adding comprehensive tests for Vagner
Instead of trying to test by effect, I'm going
to start moving in the direction of adding
a test suite for every ability on every character.
With maybe some more generic tests in some
places to cover a wider range of characters.
This should ensure that every character is always
behaving as expected, because there are a lot
of characters and abilities and I want to make
sure they all get into, and stay in, a
completely working state.