Ellya new year ability tests working
Had to do some refactoring in order to handle
more complex scenarios. Particularly around
the condition of skill activations. There are
a few combos of targets that I decided would
be easiest to just handle in a bespoke
manner for each instead of trying to figure
out some kind of logic. I'll add more
bespoke target combos as necessary to handle
more abilities.