Start testing Cipher ab3
Also a bunch of cleanup of code around
effects and conditions. Gonna be doing a
decent chunk more of that in the immediate
future in order to reduce some of the
code duplication, and to better handle
how effect targeting works.

My current next task is to figure out how
to generically handle effect targeting
properly during evaluation, because there
are some relatively-complicated scenarios
that can happen with the various combinations
that can be encountered, and they're not
actually being properly handled in the
majority of effects.